training/m2/requirements_m2.txt
pandas
numpy
scikit-learn
mlflow
pyyaml
scipy
